more admin stuff + essentials
Look, it's full of issues and stuff, I know, but hey! At least we've got some good progress and nice features now (mixed with a hell of a lot of grains of errors :,) )
Anyways here's the list of what we've been working on so far and what we're planning. Next push will probably be focusing on fixing stuff before moving on to the new features. Homepage is still not set up.... Anyways.